move modulate funcs to own file
