+++ neurodsp/tests/sim/test_modulate.py
